Add ability to shellexpand variables with their native-type values
If a variable had a non-str value like a `bool`, this would throw an
error during `shellexpand` because internally `re.sub` would try to
replace a string (the variable name) with its boolean value. Now,
with string cohercion this no longer throws an error. Additionally,
if just a single variable is specified, then the native type is kept
since `re.sub` is bypassed.