Typecast all variables
The YAML spec automatically typecasts the strings. However, if variables
are set/overwritten from the CLI or OS.ENVIRON they remain strings. This
makes for weird bugs when variables are either string or ints depending
on where they were defined.

To level the playing field, typecasting is applied to all variables.