Switch precedence for lookup of variable values
This makes it variables that are already defined, will have higher
precendence over newly defined variables. This is because if a variable
is newly defined, it is going to be ignored in this step anyway. The
only reason it is being resolved at all is just for the benefit of the
user.